src/tetra_rp/cli/templates/gpu-compute/requirements.txt
tetra-rp>=0.4.2
python-dotenv>=1.0.0
torch>=2.0.0
